task-project/task/requirements.txt
asgiref==3.9.1
attrs==25.3.0
diff-match-patch==20241021
Django==5.2.4
django-import-export==4.3.8
et_xmlfile==2.0.0
jsonrpcserver==5.0.9
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
openpyxl==3.1.5
OSlash==0.6.3
python-decouple==3.8
referencing==0.36.2
rpds-py==0.26.0
sqlparse==0.5.3
tablib==3.8.0
typing_extensions==4.14.1
tzdata==2025.2
